requirements.txt
fastapi
uvicorn
opencv-python
deepface
tf-keras
numpy
mysql-connector-python
pandas
openpyxl
jinja2
python-multipart
psutil
PyQt6 
psutil